Fix strcpy buffer overflow and mtime type

+++ src/memline.c
@@ -1091,7 +1091,7 @@ ml_recover()
     off_t	size;
     int		called_from_main;
     int		serious_error = TRUE;
-    long	mtime;
+    uint32_t	mtime;
     int		attr;
     int		orig_file_status = NOTDONE;
